+++ src/components/Table/components/index.ts
