+++ src/components/Table/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Table';
+export * from './Table';
