+++ src/components/index.ts
@@ -15,12 +15,13 @@ export { default as RadioGroup } from './RadioGroup';
 export * from './Select';
 export * from './Stepper';
 export { default as Switch } from './Switch';
-export { default as Table } from './Table';
+export * from './TableDeprecated/Table';
+export * from './Table';
 export { default as Tabs } from './Tabs';
 export { default as TextArea } from './TextArea';
 export { TextField, MuiTextField } from './TextField';
 export * from './Typography';
 
 // external components
-export {default as Divider} from '@mui/material/Divider'
-export {default as MenuItem} from '@mui/material/MenuItem'
+export { default as Divider } from '@mui/material/Divider';
+export { default as MenuItem } from '@mui/material/MenuItem';
